Cleanup kubeconfig ID
