+++ core/src/luigi-config/constants.js
@@ -25,11 +25,6 @@ export const DEFAULT_FEATURES = {
       },
     ]),
   ),
-  KUBECONFIG_ID: {
-    config: {
-      kubeconfigUrl: '/kubeconfig',
-    },
-  },
 };
 
 export const DEFAULT_HIDDEN_NAMESPACES = [
